Update sourcing-from-forestry.md
Hi! 
I think there is a mistake in this document. In the file "settings.yml" the public path should be "/uploads" and not "/static/uploads". if I've understood everything correctly, the files that live in the static folder are copied in the public directory, that means that the uploaded images will appear in "/public/uploads" and not in "/public/static/uploads". With the example settings.yml file in this doc, the images don't load, because there is no "static' directory in the public directory. But after making the change that I'm proposing here, everything works perfectly. 

I hope this helps!
Thank you for the great work!

Lucas